requirements.txt
beautifulsoup4
requests
python-dotenv
qdrant-client
tqdm
numpy
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0 